Pass resultCode to companion object
The `resultCode` parameter is now passed to the `buildCompanion` function to be used in the generated code.

The secondary constructor has been updated to avoid setting a default value for parameters already defined in `nonNullableParams`.

The logic for determining the appropriate `get...Extra` method for ArrayLists has been refactored for clarity.